Fix Block Changee deserialization

+++ src/packets/play/server/35_block_change.rs
@@ -1,9 +1,12 @@
+use gami_macros::{packet, Deserialize, Serialize};
+
 use crate::packets::Packet;
 use crate::registry::Vec3;
-use gami_macros::{packet, Deserialize, Serialize};
+use crate::serialization::{deserialize_varint, serialize_varint};
 
 #[packet(0x23, server)]
 pub struct BlockChange {
     pub location: Vec3,
-    pub block_id: u16,
+    #[encoding("varint")]
+    pub block_id: i32,
 }
